Allow barefoot players not to trample. (default disabled)
This fulfills request #4. If leather boots are soft enough to prevent
trampling crops it makes sense that barefoot players would also be
soft enough to prevent trampling. This feature is disabled by default
though because it gives players a passive buff that was only intended
for certain armor and enchantments.